Make client menu run actions
When issuing the client menu, and filtering the results, if what's entered
is not found in the menu, execute it as a command instead.

This makes the client menu work more like a launcher, without the direct
need for dmneu; furthermore, it also acts as a "swich to window if exists,
else run that command" filter, which can be handy.

This also makes the menu an enum type, which means other menu types in the
future (such as the tag menu, or rulesets menu) can have similar actions if
required.  Currently, only implemented for the client menu though.